#9 Split UserService — extract avatar retrieval into UserAvatarService
getAvatar() and its S3Repository dependency moved to new UserAvatarService.
GetAvatarAction updated to depend on UserAvatarService.
UserService retains CRUD + avatar-upload (AvatarService) logic.
